BACK/CYBEROPS-master/requirements_v3.txt
absl-py==0.7.0
aniso8601==4.1.0
astor==0.7.1
CacheControl==0.12.5
cachetools==3.1.0
cassandra-driver==3.17.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
collectd==1.0
cycler==0.10.0
dotmap==1.3.4
firebase-admin==2.16.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
future==0.17.1
gast==0.2.2
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.19.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kafka==1.3.5
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
liac-arff==2.4.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mock==2.0.0
msgpack==0.6.1
numpy==1.16.2
pandas==0.24.1
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
psutil==5.5.1
py4j==0.10.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pydub==0.23.1
pyparsing==2.3.1
pyspark==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboard==1.13.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
opencv-python
